changed int to long. added library for tests.

+++ Capstone/tenmo-client/src/main/java/com/techelevator/tenmo/services/AccountService.java
@@ -17,7 +17,7 @@ public class AccountService {
 		this.baseUrl = baseUrl;
 	}
 
-	public Account getAccount(int userId) {
+	public Account getAccount(long userId) {
 		Account account = null;
 		
 		try {
